feat: DetailViewController 추가 #1
- DetailViewController의 TextView 구현
- SplitView에 Master/Detail View 연결
- 폴더 구분
