+++ CloudNotes/.swiftlint.yml
@@ -2,6 +2,7 @@ disabled_rules: # 실행에서 제외할 룰 식별자들
   - trailing_whitespace
   - private_over_fileprivate
   - empty_enum_arguments
+  - line_length
   
 included: # 린트 과정에 포함할 파일 경로. 이 항목이 존재하면 `--path`는 무시됨
   - CloudNotes
