+++ CloudNotes/CloudNotes/View/MasterTableViewCell.swift
