--- CloudNotes/CloudNotes/ViewController.swift
@@ -1,10 +1,0 @@
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-}
